Do not mute warning 11
The warning for unused variables was muted in the case where the pattern
in the let%if binding is complete - then the catch-all pattern would
be unused and generate a warning. However, muting the warning affects
the expression as well which is undesirable, and the compiler would be
correct - you should just use let over let%if if your pattern is
complete.